Adds in basic length filtering
Adds in code that would allow the user to specify only albums
over a certain length should be added.  The code is very rough
and is commented out for the time being.  Note that the only way
currently to alter the preferred length of the value is to edit
the source code (not ideal).